% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hf_action}
\alias{hf_action}
\title{The HF-ACTION study data}
\format{
\subsection{\code{hf_action}}{

A data frame with 1,448 rows and 5 columns:
\describe{
\item{id}{The patient id}
\item{time}{Follow-up time}
\item{status}{Status: 0=censoring, 1=death, 2=hospitalization}
\item{group}{Treatment: 0=ususal care, 1=exercise training}
}
}
}
\source{
The WR package.
}
\usage{
hf_action
}
\description{
A random sample from the HF-ACTION study.
}
\keyword{datasets}
