% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R, R/winloss.R
\docType{package}
\name{winloss}
\alias{winloss}
\title{Estimate win-loss parameters and their variance}
\usage{
winloss(id, time, status, group, at = 1, type = NULL, conf.level = 0.95)
}
\arguments{
\item{id}{The id variable.}

\item{time}{The time variable.}

\item{status}{The status variable.}

\item{group}{The group variable. Assuming groups are coded 0,1.}

\item{at}{The time point where the win-loss probabilities are evaluated.}

\item{type}{The type of analysis for each event type:
1=comparing event times, 2=comparing number of recurrent events at time point at.}

\item{conf.level}{The level of the confidence interval.}
}
\value{
A list.
}
\description{
Estimate win-loss parameters and their variance.

The winloss package contain a function for inference of win-loss parameters based on censored event data.
}
\examples{
winloss(id, time, status, group, at=2)
}
\seealso{
\code{\link{winloss}} Inference
}
\concept{table functions}
